APIDJANGO/requirements.txt
pyjwt == 2.8.0
Django == 5.2.4
django-ninja==1.4.3